refactor(models/validators): extract contract violation logic to dedicated module
- Removed `ContractViolation` and related subclasses from `models.py`
- Introduced `violation_systems.py` to isolate and encapsulate violation logic
- Refactored `ModuleValidator` to delegate class validation to new `ClassValidator`
- Added `ClassValidator` for structured validation of class attributes and methods
- Improved modularity and readability of validation logic across layers

This paves the way for scalable handling of violations across contexts while keeping `models.py` focused on data structure definitions.
